(feat) add userid when sending a message through sockets
