+++ client/components/chat.js
@@ -50,7 +50,7 @@ class Chat extends React.Component {
       <div className="chatBox">
         <table className="table-striped">
           <tbody>
-            {messages.map(message => <Message key={message.id} user={message.user} text={message.text} timestamp={message.createdAt}/>)}
+            {messages.map(message => <Message key={message.id} userid={message.userid} user={message.user} text={message.text} timestamp={message.createdAt}/>)}
           </tbody>
         </table>
         <Input />
